merge new version

+++ heron/statemgrs/src/java/com/twitter/heron/statemgr/localfs/LocalFileSystemStateManager.java
@@ -14,7 +14,9 @@
 
 package com.twitter.heron.statemgr.localfs;
 
-import java.util.logging.Level;
+import java.nio.charset.Charset;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
 import com.google.common.util.concurrent.ListenableFuture;
@@ -26,25 +28,61 @@ import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.common.basics.FileUtils;
 import com.twitter.heron.proto.scheduler.Scheduler;
 import com.twitter.heron.proto.system.ExecutionEnvironment;
+import com.twitter.heron.proto.system.PackingPlans;
 import com.twitter.heron.proto.system.PhysicalPlans;
 import com.twitter.heron.proto.tmaster.TopologyMaster;
 import com.twitter.heron.spi.common.Config;
+import com.twitter.heron.spi.common.Keys;
+import com.twitter.heron.spi.statemgr.Lock;
 import com.twitter.heron.spi.statemgr.WatchCallback;
 import com.twitter.heron.statemgr.FileSystemStateManager;
 
 public class LocalFileSystemStateManager extends FileSystemStateManager {
   private static final Logger LOG = Logger.getLogger(LocalFileSystemStateManager.class.getName());
 
-  private Config config;
+  /**
+   * Local filesystem implementation of a lock that mimics the file system behavior of the
+   * distributed lock.
+   */
+  private final class FileSystemLock implements Lock {
+    private String path;
+
+    private FileSystemLock(String path) {
+      this.path = path;
+    }
+
+    @Override
+    public boolean tryLock(long timeout, TimeUnit unit) throws InterruptedException {
+      long giveUpAtMillis = System.currentTimeMillis() + unit.toMillis(timeout);
+      byte[] fileContents = Thread.currentThread().getName().getBytes(Charset.defaultCharset());
+      while (true) {
+        try {
+          if (setData(this.path, fileContents, false).get()) {
+            return true;
+          } else if (System.currentTimeMillis() >= giveUpAtMillis) {
+            return false;
+          } else {
+            TimeUnit.SECONDS.sleep(2); // need to pole the filesystem for availability
+          }
+        } catch (ExecutionException e) {
+          // this is thrown when the file exists, which means the lock can't be obtained
+        }
+      }
+    }
+
+    @Override
+    public void unlock() {
+      deleteNode(this.path, false);
+    }
+  }
 
   @Override
   public void initialize(Config ipconfig) {
 
     super.initialize(ipconfig);
-    this.config = ipconfig;
 
     // By default, we would init the file tree if it is not there
-    boolean isInitLocalFileTree = LocalFileSystemContext.initLocalFileTree(config);
+    boolean isInitLocalFileTree = LocalFileSystemContext.initLocalFileTree(ipconfig);
 
     if (isInitLocalFileTree && !initTree()) {
       throw new IllegalArgumentException("Failed to initialize Local State manager. "
@@ -53,34 +91,15 @@ public class LocalFileSystemStateManager extends FileSystemStateManager {
   }
 
   protected boolean initTree() {
-    // Make necessary directories
-    LOG.log(Level.FINE, "Topologies directory: {0}", getTopologyDir());
-    LOG.log(Level.FINE, "Tmaster location directory: {0}", getTMasterLocationDir());
-    LOG.log(Level.FINE, "Physical plan directory: {0}", getPhysicalPlanDir());
-    LOG.log(Level.FINE, "Execution state directory: {0}", getExecutionStateDir());
-    LOG.log(Level.FINE, "Scheduler location directory: {0}", getSchedulerLocationDir());
-
-    boolean topologyDir = FileUtils.isDirectoryExists(getTopologyDir())
-        || FileUtils.createDirectory(getTopologyDir());
-
-    boolean tmasterLocationDir = FileUtils.isDirectoryExists(getTMasterLocationDir())
-        || FileUtils.createDirectory(getTMasterLocationDir());
-
-    boolean physicalPlanDir = FileUtils.isDirectoryExists(getPhysicalPlanDir())
-        || FileUtils.createDirectory(getPhysicalPlanDir());
-
-    boolean executionStateDir = FileUtils.isDirectoryExists(getExecutionStateDir())
-        || FileUtils.createDirectory(getExecutionStateDir());
-
-    boolean schedulerLocationDir = FileUtils.isDirectoryExists(getSchedulerLocationDir())
-        || FileUtils.createDirectory(getSchedulerLocationDir());
-
-    if (topologyDir && tmasterLocationDir && physicalPlanDir && executionStateDir
-        && schedulerLocationDir) {
-      return true;
+    for (StateLocation location : StateLocation.values()) {
+      String dir = getStateDirectory(location);
+      LOG.fine(String.format("%s directory: %s", location.getName(), dir));
+      if (!FileUtils.isDirectoryExists(dir) && !FileUtils.createDirectory(dir)) {
+        return false;
+      }
     }
 
-    return false;
+    return true;
   }
 
   // Make utils class protected for easy unit testing
@@ -92,18 +111,50 @@ public class LocalFileSystemStateManager extends FileSystemStateManager {
     return future;
   }
 
-  protected ListenableFuture<Boolean> deleteData(String path) {
-    final SettableFuture<Boolean> future = SettableFuture.create();
-    boolean ret = FileUtils.deleteFile(path);
+  private ListenableFuture<Boolean> setData(StateLocation location,
+                                            String topologyName,
+                                            byte[] bytes,
+                                            boolean overwrite) {
+    return setData(getStatePath(location, topologyName), bytes, overwrite);
+  }
+
+  @Override
+  protected ListenableFuture<Boolean> nodeExists(String path) {
+    SettableFuture<Boolean> future = SettableFuture.create();
+    boolean ret = FileUtils.isFileExists(path);
     future.set(ret);
 
     return future;
   }
 
+  @Override
+  protected ListenableFuture<Boolean> deleteNode(String path, boolean deleteChildrenIfNecessary) {
+    final SettableFuture<Boolean> future = SettableFuture.create();
+    boolean ret = true;
+    if (FileUtils.isFileExists(path)) {
+      if (!deleteChildrenIfNecessary && FileUtils.hasChildren(path)) {
+        LOG.severe("delete called on a path with children but deleteChildrenIfNecessary is false: "
+            + path);
+        ret = false;
+      } else {
+        ret = FileUtils.deleteFile(path);
+      }
+    }
+    future.set(ret);
+
+    return future;
+  }
+
+  @Override
   @SuppressWarnings("unchecked") // we don't know what M is until runtime
-  protected <M extends Message> ListenableFuture<M> getData(String path, Message.Builder builder) {
+  protected <M extends Message> ListenableFuture<M> getNodeData(WatchCallback watcher,
+                                                                String path,
+                                                                Message.Builder builder) {
     final SettableFuture<M> future = SettableFuture.create();
-    byte[] data = FileUtils.readFromFile(path);
+    byte[] data = new byte[]{};
+    if (FileUtils.isFileExists(path)) {
+      data = FileUtils.readFromFile(path);
+    }
     if (data.length == 0) {
       future.set(null);
       return future;
@@ -122,7 +173,8 @@ public class LocalFileSystemStateManager extends FileSystemStateManager {
   @Override
   public ListenableFuture<Boolean> setExecutionState(
       ExecutionEnvironment.ExecutionState executionState, String topologyName) {
-    return setData(getExecutionStatePath(topologyName), executionState.toByteArray(), false);
+    return setData(
+        StateLocation.EXECUTION_STATE, topologyName, executionState.toByteArray(), false);
   }
 
   @Override
@@ -131,18 +183,24 @@ public class LocalFileSystemStateManager extends FileSystemStateManager {
     // Note: Unlike Zk statemgr, we overwrite the location even if there is already one.
     // This is because when running in simulator we control when a tmaster dies and
     // comes up deterministically.
-    return setData(getTMasterLocationPath(topologyName), location.toByteArray(), true);
+    return setData(StateLocation.TMASTER_LOCATION, topologyName, location.toByteArray(), true);
   }
 
   @Override
   public ListenableFuture<Boolean> setTopology(TopologyAPI.Topology topology, String topologyName) {
-    return setData(getTopologyPath(topologyName), topology.toByteArray(), false);
+    return setData(StateLocation.TOPOLOGY, topologyName, topology.toByteArray(), false);
   }
 
   @Override
   public ListenableFuture<Boolean> setPhysicalPlan(
       PhysicalPlans.PhysicalPlan physicalPlan, String topologyName) {
-    return setData(getPhysicalPlanPath(topologyName), physicalPlan.toByteArray(), false);
+    return setData(StateLocation.PHYSICAL_PLAN, topologyName, physicalPlan.toByteArray(), false);
+  }
+
+  @Override
+  public ListenableFuture<Boolean> setPackingPlan(
+      PackingPlans.PackingPlan packingPlan, String topologyName) {
+    return setData(StateLocation.PACKING_PLAN, topologyName, packingPlan.toByteArray(), true);
   }
 
   @Override
@@ -151,80 +209,27 @@ public class LocalFileSystemStateManager extends FileSystemStateManager {
     // Note: Unlike Zk statemgr, we overwrite the location even if there is already one.
     // This is because when running in simulator we control when a scheduler dies and
     // comes up deterministically.
-    return setData(getSchedulerLocationPath(topologyName), location.toByteArray(), true);
+    return setData(StateLocation.SCHEDULER_LOCATION, topologyName, location.toByteArray(), true);
   }
 
   @Override
-  public ListenableFuture<Boolean> deleteTMasterLocation(String topologyName) {
-    return deleteData(getTMasterLocationPath(topologyName));
-  }
-
-  @Override
-  public ListenableFuture<Boolean> deleteSchedulerLocation(String topologyName) {
-    return deleteData(getSchedulerLocationPath(topologyName));
-  }
-
-  @Override
-  public ListenableFuture<Boolean> deleteExecutionState(String topologyName) {
-    return deleteData(getExecutionStatePath(topologyName));
-  }
-
-  @Override
-  public ListenableFuture<Boolean> deleteTopology(String topologyName) {
-    return deleteData(getTopologyPath(topologyName));
-  }
-
-  @Override
-  public ListenableFuture<Boolean> deletePhysicalPlan(String topologyName) {
-    return deleteData(getPhysicalPlanPath(topologyName));
-  }
-
-  @Override
-  public ListenableFuture<Scheduler.SchedulerLocation> getSchedulerLocation(
-      WatchCallback watcher, String topologyName) {
-    return getData(getSchedulerLocationPath(topologyName),
-        Scheduler.SchedulerLocation.newBuilder());
-  }
-
-  @Override
-  public ListenableFuture<TopologyAPI.Topology> getTopology(
-      WatchCallback watcher, String topologyName) {
-    return getData(getTopologyPath(topologyName), TopologyAPI.Topology.newBuilder());
-  }
-
-  @Override
-  public ListenableFuture<ExecutionEnvironment.ExecutionState> getExecutionState(
-      WatchCallback watcher, String topologyName) {
-    return getData(getExecutionStatePath(topologyName),
-        ExecutionEnvironment.ExecutionState.newBuilder());
-  }
-
-  @Override
-  public ListenableFuture<PhysicalPlans.PhysicalPlan> getPhysicalPlan(
-      WatchCallback watcher, String topologyName) {
-    return getData(getPhysicalPlanPath(topologyName),
-        PhysicalPlans.PhysicalPlan.newBuilder());
-  }
-
-  @Override
-  public ListenableFuture<TopologyMaster.TMasterLocation> getTMasterLocation(
-      WatchCallback watcher, String topologyName) {
-    return getData(getTMasterLocationPath(topologyName),
-        TopologyMaster.TMasterLocation.newBuilder());
-  }
-
-  @Override
-  public ListenableFuture<Boolean> isTopologyRunning(String topologyName) {
-    SettableFuture<Boolean> future = SettableFuture.create();
-    boolean ret = FileUtils.isFileExists(getTopologyPath(topologyName));
-    future.set(ret);
-
-    return future;
+  protected Lock getLock(String path) {
+    return new FileSystemLock(path);
   }
 
   @Override
   public void close() {
     // We would not clear anything here
     // Scheduler kill interface should take care of the cleaning
+  }
+
+  public static void main(String[] args) throws ExecutionException, InterruptedException,
+      IllegalAccessException, ClassNotFoundException, InstantiationException {
+    Config config = Config.newBuilder()
+        .put(Keys.stateManagerRootPath(),
+            System.getProperty("user.home") + "/.herondata/repository/state/local")
+        .build();
+    LocalFileSystemStateManager stateManager = new LocalFileSystemStateManager();
+    stateManager.doMain(args, config);
   }
 }
